프로젝트 파일 정리 및 플레이어 상호작용 활성화 시, 움직이지 않도록 변경

+++ AvoidTheBoss/Scene.h
@@ -3,6 +3,9 @@
 #include "SceneInterface.h"
 #include "Shader.h"
 #include "Player.h"
+#include "CEmployee.h"
+#include "CBoss.h"
+
 
 #define MAX_LIGHTS			16 
 
@@ -84,7 +87,7 @@ public:
 
 	CPlayer*					_players[4];
 	int16						_playerIdx = 0;
-	DWORD						m_lastKeyInput = 0;
+	uint8						m_lastKeyInput = 0;
 	//마지막으로 마우스 버튼을 클릭할 때의 마우스 커서의 위치이다. 
 	POINT								m_ptOldCursorPos;
